Maru.Eval.RuntimeOperation

Description

Define functions and macros, these are used in the runtime,

also these are through with MaruEnv (it should be initialEnv)

These respects clisp's behavior basically.

add, sub, times, and div are regarded as axiomly functions.

Documentation

add :: MaruFunc

>>> runMaruCalculator $ execFunc add [AtomInt 1, AtomInt 2]
Right (AtomInt 3)
>>> runMaruCalculator $ execFunc add []
Right (AtomInt 0)
>>> isLeft . runMaruCalculator $ execFunc add [AtomSymbol "xD"]
True

sub :: MaruFunc

>>> runMaruCalculator $ execFunc sub [AtomInt 3, AtomInt 1]
Right (AtomInt 2)
>>> runMaruCalculator $ execFunc sub [AtomInt 1]
Right (AtomInt (-1))
>>> isLeft . runMaruCalculator $ execFunc sub []
True
>>> isLeft . runMaruCalculator $ execFunc sub [AtomSymbol "xD"]
True

times :: MaruFunc

>>> runMaruCalculator $ execFunc times [AtomInt 3, AtomInt 3]
Right (AtomInt 9)
>>> runMaruCalculator $ execFunc times []
Right (AtomInt 1)
>>> isLeft . runMaruCalculator $ execFunc times [AtomSymbol "xD"]
True

div :: MaruFunc

>>> runMaruCalculator $ execFunc div [AtomInt 3, AtomInt 3]
Right (AtomInt 1)
>>> isLeft . runMaruCalculator $ execFunc div []
True
>>> isLeft . runMaruCalculator $ execFunc div [AtomSymbol "xD"]
True
>>> isLeft . runMaruCalculator $ execFunc div [AtomInt 0, AtomInt 1]
True
>>> runMaruCalculator $ execFunc div [AtomInt 10, AtomInt 3]
Right (AtomInt 3)
>>> runMaruCalculator $ execFunc div [AtomInt 3, AtomInt 5]
Right (AtomInt 0)