Maru.Type.Eval

Description

Integrate types of extensible's Effect.

MaruMacro is evaluated by MaruEvaluator. MaruFunc is calculated by MaruCalculator.

Documentation

type Fail = FailKey >: FailValue

An effect of MaruEvaluator. A possible of the failure.

type FailKey = "fail"

type FailValue = EitherEff ExceptionCause

type FailAssociation = Associate FailKey FailValue

Fail's Associate

throwFail :: FailAssociation xs => ExceptionCause -> Eff xs a

throwEff for Fail

includeFail :: FailAssociation xs => ExceptionCause -> Eff xs (Maybe a) -> Eff xs a

Include Maybe to Fail context. If it is Nothing, the whole of `Fail a` to be failed.

includeFail' :: FailAssociation xs => Eff xs (Either ExceptionCause a) -> Eff xs a

Same as includeFail

type SimplificationSteps = [SExpr]

A log for 簡約s

reportSteps :: SimplificationSteps -> [Text]

Append numbers to steps

>>> import Maru.Type.SExpr
>>> reportSteps [Cons (AtomInt 1) (Cons (AtomInt 2) Nil), Cons (AtomInt 2) Nil]
["1: (1 2)","2: (2)"]

type SimplifSteps = SimplifStepsKey >: SimplifStepsValue

An effect of MaruEvaluator, for logging simplifications

type SimplifStepsKey = "simplifSteps"

type SimplifStepsValue = WriterEff SimplificationSteps

type SimplifStepsAssociation = Associate SimplifStepsKey SimplifStepsValue

type MaruScopes = MaruScopesKey >: MaruScopesValue

An effect of MaruEvaluator. This is a stack for the lexical scope.

type MaruScopesKey = "maruScopes"

type MaruScopesValue = State MaruEnv

type MaruScopesAssociation = Associate MaruScopesKey MaruScopesValue

insertGlobalVar :: MaruScopesAssociation xs => MaruSymbol -> SExpr -> Eff xs ()

Insert a variable to the toplevel scope

>>> (_, env, _) <- flip runMaruEvaluator E.initialEnv $ insertGlobalVar "x" (AtomInt 10)
>>> lookup "x" env
Just (AtomInt 10)

newScope :: MaruScopesAssociation xs => MaruSymbol -> SExpr -> Eff xs ()

Make a scope in the state, with a variable.

NOTE: The scope must be created with one or more variables. It keeps any safety.

e.g. 1. unintended empty scope is never created 2. high affinity of MaruEnv (NonEmpty) is kept

popNewerScope :: MaruScopesAssociation xs => Eff xs MaruScope

Remove the newest scope (about the newest scope is written in MaruEnv), and Return removed scope

type MaruEnv = NonEmpty MaruScope

The whole of the runtime state. This is NonEmpty, because the global scope is defined with the program startup

and

'getMaruEnv >>= return . last' is the toplevel (and this is used as global scope). 'getMaruEnv >>= return . head' is the newest scope.

( This means the cons operation makes a new scope, it is not '++ [x]'. Please see makeScope. )

getMaruEnv :: MaruScopesAssociation xs => Eff xs MaruEnv

getEff for MaruScopes

putMaruEnv :: MaruScopesAssociation xs => MaruEnv -> Eff xs ()

putEff for MaruScopes

modifyMaruEnv :: MaruScopesAssociation xs => (MaruEnv -> MaruEnv) -> Eff xs ()

modifyEff for maruScopes

type IOEff = IOEffKey >: IOEffValue

An effect of MaruEvaluator, this is same as IO in Eff

type IOEffKey = "ioEff"

type IOEffValue = IO

type IOEffAssociation = Associate IOEffKey IOEffValue

liftIOEff :: IOEffAssociation xs => IO a -> Eff xs a

liftEff for IOEff

type ExceptionCause = Text

A message of Fail

type MaruEvaluator = Eff '[Fail, MaruScopes, SimplifSteps, IOEff]

A monad for evaluating a maru's program

runMaruEvaluator :: MaruEvaluator a -> MaruEnv -> IO (Either ExceptionCause a, MaruEnv, SimplificationSteps)

Run an evaluation of MaruEvaluator a

newSymbol :: MaruEvaluator MaruSymbol

Create a symbol of the variable name, it is unique (is not duplicated) in the runtime.

type MaruScope = Map MaruSymbol SExpr

The runtime state. This associates the variable name and the value.

newtype MaruFunc

A function of maru. This keeps the purity, don't happen effects.

Take [SExpr] as arguments, its length is checked by each function. If it is not the expected length, Nothing maybe given.

Notice:

The function is Haskell's function, is represented by Haskell. The function is not maru's (runtime's) function (cannot be defined in the runtime).

Constructors

MaruFunc

Fields

execFunc :: [SExpr] -> MaruCalculator SExpr

newtype MaruMacro

A macro of maru, this means the impure function.

Similar to MaruFunc, but this is possibility to update the state of the environment.

Constructors

MaruMacro

Fields

execMacro :: SExpr -> MaruEvaluator SExpr

lookup :: MaruSymbol -> MaruEnv -> Maybe SExpr

Find a variable from the whole of the runtime environment with a symbol

lookupVar :: forall xs. (FailAssociation xs, MaruScopesAssociation xs) => MaruSymbol -> Eff xs SExpr

Take a variable from MaruScopes effect. If sym is not exists, the whole of this Eff to be failed

(^$) :: (Typeable s, Typeable a) => MaruEvaluator s -> Getting (First a) s a -> MaruEvaluator a

This is like Prism's accessor, but don't return result as Maybe.

Similar to 'x & review f' but Nothing is included as a failure of the whole of MaruEvaluator.

Typeable for the error message.

type MaruCalculator = Eff '[Fail]

This has effects of the part of MaruEvaluator. Calculate pure functions.

runMaruCalculator :: MaruCalculator a -> Either ExceptionCause a

Extract the pure calculation from MaruCalculator

newtype First' a

Simular to First, but using 'Either ExceptionCause' instead of Maybe

Constructors

First'

Fields

getFirst' :: Either ExceptionCause a

first' :: Maybe a -> First' a

Like a consturctor, but from 'Maybe a'. If Nothing is given, return mempty.

expandVars :: (MaruScopesAssociation xs, FailAssociation xs) => SExpr -> Eff xs SExpr

Expand the value of the variables, but these are not evaluated.

And +, -, *, and / are not expanded (because it is regarded as like the axioms)

simply expanding

>>> (sexpr, _, _) <- flip runMaruEvaluator E.initialEnv $ newScope "x" (AtomInt 10) >> expandVars (AtomSymbol "x")
>>> sexpr
Right (AtomInt 10)

multi variables

>>> (sexpr, _, _) <- flip runMaruEvaluator E.initialEnv $ newScope "x" (AtomInt 10) >> newScope "y" (AtomBool True) >> expandVars (Cons (AtomSymbol "x") (Cons (AtomSymbol "y") Nil))
>>> sexpr
Right (Cons (AtomInt 10) (Cons (AtomBool True) Nil))

nested expanding

>>> (sexpr, _, _) <- flip runMaruEvaluator E.initialEnv $ newScope "x" (AtomInt 10) >> newScope "y" (AtomSymbol "x") >> expandVars (AtomSymbol "y")
>>> sexpr
Right (AtomInt 10)

substituteVar :: MaruSymbol -> SExpr -> SExpr -> SExpr

Substitute a value to a variable.

"x" is substituted by `AtomInt 10` in `Cons (AtomInt 1) (Cons (AtomSymbol "x") Nil)`.

>>> substituteVar "x" (AtomInt 10) $ Cons (AtomInt 1) (Cons (AtomSymbol "x") Nil)
Cons (AtomInt 1) (Cons (AtomInt 10) Nil)

>>> substituteVar "x" (AtomInt 10) $ Cons (AtomSymbol "x") (Cons (AtomSymbol "x") Nil)
Cons (AtomInt 10) (Cons (AtomInt 10) Nil)

Orphan instances

MonadFail MaruCalculator

Methods

fail :: String -> MaruCalculator a

MonadFail MaruEvaluator

Methods

fail :: String -> MaruEvaluator a