Maru.Type.SExpr

Description

Common types for zuramaru

Documentation

type SourceCode = Text

The format for the code of maru. (This doesn't mean a file path of the code.)

type MaruToken = Token Text

The format for the token of MaruParser

data SExpr

n-ary tree and terms

Constructors

Cons SExpr SExpr
Appending list and list

Nil
A representation of empty list

AtomInt Int
A pattern of the atom for Int

AtomBool Bool
A pattern of the atom for Bool

AtomSymbol MaruSymbol
A pattern of the atom for MaruSymbol

isAtomInt :: SExpr -> Bool

>>> isAtomInt $ AtomInt 10
True
>>> isAtomInt Nil
False
>>> isAtomInt $ AtomSymbol ""
False

unAtomInt :: SExpr -> Maybe Int

Extract Int from a term of AtomInt.

>>> unAtomInt $ AtomInt 10
Just 10
>>> unAtomInt $ AtomSymbol ":D"
Nothing
>>> unAtomInt $ Cons (AtomInt 10) (AtomInt 20)
Nothing

isAtomSymbol :: SExpr -> Bool

>>> isAtomSymbol $ AtomSymbol "x"
True
>>> isAtomSymbol Nil
False
>>> isAtomSymbol $ AtomInt 10
False

unAtomSymbol :: SExpr -> Maybe MaruSymbol

Similar to unAtomInt

class SExprLike a where

a can be represented as SExpr

Minimal complete definition

wrap

Methods

wrap :: a -> SExpr

a can be converted as SExpr

readable :: SExpr -> Text

Show SExpr as the human readable syntax. This is the inverse function of the parser, if the format is ignored (e.g. '( + 1 2)` =~ '(+ 1 2)').

>>> readable <$> parse "10"
Right "10"
>>> parse . readable $ AtomInt 10
Right (AtomInt 10)

>>> readable <$> parse "true"
Right "true"
>>> parse . readable $ AtomBool True
Right (AtomBool True)

>>> readable <$> parse "(+ 1 2)"
Right "(+ 1 2)"
>>> let result = parse . readable $ Cons (AtomSymbol "+") (Cons (AtomInt 1) (Cons (AtomInt 2) Nil))
>>> result == Right (Cons (AtomSymbol "+") (Cons (AtomInt 1) (Cons (AtomInt 2) Nil)))
True

newtype MaruSymbol

A symbol of MaruEnv, but this is not meaning a symbol of maru side

Constructors

MaruSymbol

Fields

unMaruSymbol :: Text

pack :: String -> MaruSymbol

Wrap String. If you want to wrap Text, please use MaruSymbol value constructor instead.

unpack :: MaruSymbol -> String

A dual of pack

asSymbolList :: Prism' [SExpr] [MaruSymbol]

A Prism accessor.

Get Nothing if [SExpr] includes non AtomSymbol. Get all AtomSymbol otherwise.

>>> [AtomSymbol "x", AtomSymbol "y"] ^? asSymbolList
Just ["x","y"]

>>> [AtomInt 1, AtomSymbol "y"] ^? asSymbolList
Nothing

scottEncode :: [SExpr] -> SExpr

Concatenate SExpr by Cons

>>> let xs = [(AtomInt 1), (Cons (Cons (AtomInt 2) (Cons (AtomInt 3) Nil)) Nil)] -- [1, (2 3)]
>>> scottEncode xs
Cons (AtomInt 1) (Cons (Cons (Cons (AtomInt 2) (Cons (AtomInt 3) Nil)) Nil) Nil)
>>> let ys = [AtomInt 1, AtomInt 2, AtomInt 3] -- [1, 2, 3]
>>> scottEncode ys
Cons (AtomInt 1) (Cons (AtomInt 2) (Cons (AtomInt 3) Nil))
>>> let zs = [AtomInt 1, Nil] -- [1, ()]
>>> scottEncode zs
Cons (AtomInt 1) (Cons Nil Nil)

scottDecode :: SExpr -> [SExpr]

The inverse function of scottEncode

>>> let xs = Cons (AtomInt 1) (Cons (AtomInt 2) Nil)
>>> scottDecode xs
[AtomInt 1,AtomInt 2]
>>> scottDecode $ Cons (AtomInt 10) Nil
[AtomInt 10]
>>> scottDecode $ Cons Nil Nil
[Nil]

_Cons :: Prism' SExpr (SExpr, SExpr)

_Nil :: Prism' SExpr ()

_AtomInt :: Prism' SExpr Int

_AtomSymbol :: Prism' SExpr MaruSymbol

newtype SExprIntBullet

Shot only the AtomInts by omap

Constructors

SExprIntBullet

Fields

unSExprIntBullet :: SExpr

intBullet :: (Int -> Int) -> SExpr -> SExpr

Apply by omap a function to a SExprIntBullet with wrapping and unwrapping